Support serializing models larger than 2**31 - 1 (#624)
* Support serializing models larger than 2**31 - 1

Previously `treelite` used `ctypes.string_at` to copy the serialized
bytes return value to a new python `bytes` object. This method takes a
pointer and a length (expressed as an `int`). Python `bytes` objects
have a max capacity of `Py_ssize_t`, not `int`. This meant that
serializing very large models could error as the `size` parameter would
overflow an `int`.

We now use `PyBytes_FromStringAndSize` directly, avoiding this issue.

It's hard to write a sane test for this that can run on CI, but I've
verified that things are working locally.

* Fix tests

---------

Co-authored-by: Philip Hyunsu Cho <chohyu01@cs.washington.edu>
Co-authored-by: Hyunsu Cho <phcho@nvidia.com>